+++ src/resolvers/conferenceResolvers.ts
@@ -1,0 +1,7 @@
+export default {
+  Query: {
+    conferences: (_source: any, _: any, { dataSources }: any) => {
+      return dataSources.conferencesApi.getConferences();
+    },
+  },
+};
