+++ src/resolvers/index.ts
@@ -1,0 +1,3 @@
+import conferenceResolvers from './conferenceResolvers';
+
+export default [conferenceResolvers];
