+++ src/typeDefs/index.ts
@@ -1,0 +1,4 @@
+import query from "./query";
+import types from "./types";
+
+export default [query, types];
