+++ src/typeDefs/query.ts
@@ -1,0 +1,7 @@
+import { gql } from 'apollo-server';
+
+export default gql`
+  type Query {
+    conferences: [Conference]
+  }
+`;
