+++ src/typeDefs/types/conferenceType.ts
@@ -1,0 +1,8 @@
+import { gql } from 'apollo-server';
+
+export default gql`
+  type Conference {
+    uniqueName: String
+    displayName: String
+  }
+`;
