+++ src/typeDefs/types/index.ts
@@ -1,0 +1,3 @@
+import conferenceType from "./conferenceType";
+
+export default conferenceType;
