+++ tsconfig.json
@@ -1,7 +1,7 @@
 {
     "compilerOptions": {
-        "module": "esnext",
-        "target": "esnext",
+        "module": "commonjs",
+        "target": "es6",
         "sourceMap": true,
         "allowJs": false,
         "noImplicitReturns": true,
@@ -16,7 +16,8 @@
         "noFallthroughCasesInSwitch": true,
         "allowSyntheticDefaultImports": false,
         "removeComments": true,
-        "skipLibCheck": true
+        "skipLibCheck": true,
+        "moduleResolution": "node"
     },
     "include": [
         "src/**/*"
